% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_transcriptID.R
\name{convert_transcriptID}
\alias{convert_transcriptID}
\title{Convert transcript IDs between different databases}
\usage{
convert_transcriptID(
  dat,
  db,
  biomart = "ensembl",
  dataset = "hsapiens_gene_ensembl",
  getBM_attributes = c("refseq_mrna", "ensembl_transcript_id", "hgnc_symbol"),
  filters = "refseq_mrna"
)
}
\arguments{
\item{dat}{a dataframe including Transcript_version, nucleotide "C1768G"
refers to ref/CDS position/alt.}

\item{db}{`EnsDb` object. Default is EnsDb.Hsapiens.v75.}

\item{biomart}{selection of BioMart database. Default is "ensembl".}

\item{dataset}{BioMart databases includes many datasets. Choose dataset in the database.
Default is "hsapiens_gene_ensembl".}

\item{getBM_attributes}{defines the values of interests.
Default shows "refseq_mrna", "ensembl_transcript_id", "hgnc_symbol".
The listAttributes function displays all available attributes in the selected dataset.}

\item{filters}{a vector of filters to query. Default shows "refseq_mrna".}
}
\value{
a new dataset with converting information
}
\description{
Preprocessing, converting and mapping
}
\examples{
library(EnsDb.Hsapiens.v75)
db=EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75
dat<-read.csv(system.file("extdata",
                          "convertID_data.csv",
                          package = "MRAT"),
              stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
new_dat<-convert_transcriptID(dat, db, biomart, dataset, getBM_attributes, filters)

}
