% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-convertID_result.R
\name{convertID_result}
\alias{convertID_result}
\title{Convert transcript IDs between different databases}
\format{
A data frame with 3 rows and 17 variables:
\describe{
  \item{\code{refseq}}{character COLUMN_DESCRIPTION}
  \item{\code{Nucleotide_changes}}{character COLUMN_DESCRIPTION}
  \item{\code{CDS_start_loc}}{integer COLUMN_DESCRIPTION}
  \item{\code{ref}}{character COLUMN_DESCRIPTION}
  \item{\code{alt}}{character COLUMN_DESCRIPTION}
  \item{\code{chr}}{integer COLUMN_DESCRIPTION}
  \item{\code{ensembl_transcript_id}}{character COLUMN_DESCRIPTION}
  \item{\code{hgnc_symbol}}{character COLUMN_DESCRIPTION}
  \item{\code{start}}{integer COLUMN_DESCRIPTION}
  \item{\code{end}}{integer COLUMN_DESCRIPTION}
  \item{\code{width}}{integer COLUMN_DESCRIPTION}
  \item{\code{strand}}{integer COLUMN_DESCRIPTION}
  \item{\code{exon_id}}{character COLUMN_DESCRIPTION}
  \item{\code{exon_rank}}{integer COLUMN_DESCRIPTION}
  \item{\code{tx_start}}{integer COLUMN_DESCRIPTION}
  \item{\code{tx_end}}{integer COLUMN_DESCRIPTION}
  \item{\code{refsnp_id}}{character COLUMN_DESCRIPTION}
}
}
\description{
Preprocessing, converting and mapping
}
