% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_freq.R
\name{pop_freq}
\alias{pop_freq}
\title{search allel frequency among populations and gene annotation information}
\usage{
pop_freq(ori_data, pop = "db_gnomAD_exome_freq")
}
\arguments{
\item{pop}{selection of table. Default is db_TWB_GWG_freq}

\item{dat_ori}{a dataframe including chromosome, start and end postion, ref/alt nucleotide}
}
\value{
a new dataset with converting information
}
\description{
using genomic postion and ref/alt information to query gene annotation and allele frequency
}
\examples{
dat<-read.csv(system.file("extdata",
                          "anno_freq_data.csv",
                          package = "MRAT"),
              stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
pop_dat<-pop_freq(dat, pop)

}
