% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_transcriptID.R
\name{convert_transcriptID}
\alias{convert_transcriptID}
\title{Convert transcript IDs between different databases}
\usage{
convert_transcriptID(
  dat,
  db,
  biomart_ens = "ensembl",
  dat_ens = "hsapiens_gene_ensembl",
  dat_filter = "refseq_mrna",
  BM_att_ens = c("refseq_mrna", "ensembl_transcript_id", "hgnc_symbol", "ucsc")
)
}
\arguments{
\item{dat}{a dataframe including Transcript_version, nucleotide "C1768G"
refers to ref/CDS position/alt.}

\item{db}{`EnsDb` object. Default is EnsDb.Hsapiens.v75.}

\item{biomart_ens}{selection of BioMart database. Default is "ensembl".}

\item{dat_ens}{BioMart databases includes many datasets. Choose dataset in the database.
Default is "hsapiens_gene_ensembl". Alternatives include "drerio_gene_ensembl" and "mmusculus_gene_ensembl".}

\item{dat_filter}{refers to the types of transcripts. Default is "refseq_mrna". The user can also specify "ucsc" as transcript ID.}

\item{BM_att_ens}{defines the values of interests.
Default shows "refseq_mrna", "ensembl_transcript_id", "hgnc_symbol", "ucsc".
The listAttributes function displays all available attributes in the selected dataset.}
}
\value{
a new dataset with converting information
}
\description{
Preprocessing, converting and mapping
}
\examples{
library(EnsDb.Hsapiens.v75)
db=EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75
dat<-read.csv(system.file("extdata",
                          "convertID_refseq_data.csv",
                          package = "TransAT"),
              stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
new_dat<-convert_transcriptID(dat, db, dat_filter = "refseq_mrna")

}
