% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_freq.R
\name{pop_freq}
\alias{pop_freq}
\title{search allele frequency among populations and gene annotation information}
\usage{
pop_freq(dat_ori, pop = "db_gnomAD_exome_freq")
}
\arguments{
\item{dat_ori}{a dataframe including chromosome, start and end postion, ref/alt nucleotide}

\item{pop}{selection of table. Default is db_gnomAD_exome_freq}
}
\value{
a new dataset with allele frequency and annotation information
}
\description{
using genomic postion and ref/alt information to query gene annotation and allele frequency
}
\examples{
dat<-read.csv(system.file("extdata",
                          "anno_freq_data.csv",
                          package = "TransAT"),
              stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
pop_dat<-pop_freq(dat, pop="db_gnomAD_exome_freq")

}
